% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_s_hwci.R
\name{calculate_s_hwci}
\alias{calculate_s_hwci}
\title{Title}
\usage{
calculate_s_hwci(
  t,
  j,
  v,
  m,
  b,
  l,
  w,
  r,
  s = c("LC", "NT", "VU", "EN", "CR", "NE", "DD")
)
}
\arguments{
\item{s}{}
}
\description{
Title
}
\author{
geryan
}
