% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwci.R
\name{hwci}
\alias{hwci}
\title{HWC Index suite}
\usage{
hwci(data)
}
\arguments{
\item{data}{}
}
\value{
A \code{hwciresults} object containing the HWC Index suite.
}
\description{
Calculate HWC Index suite for single species or landscape aggregate
}
\author{
Gerard Ryan
}
