% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwci.tools-package.R
\docType{package}
\name{hwci.tools-package}
\alias{hwci.tools}
\alias{hwci.tools-package}
\title{hwci.tools: Tools for Calculating an Index of Human-Wildlife Conflict}
\description{
Tools for calculating an index of human-wildlife conflict.
}
\author{
\strong{Maintainer}: Gerard Ryan \email{gerard.ryan@telethonkids.org.au} (\href{https://orcid.org/0000-0003-0183-7630}{ORCID})

}
\keyword{internal}
