% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwci_data.R
\name{hwci_data}
\alias{hwci_data}
\title{HWC Index data}
\usage{
hwci_data(
  t,
  j,
  v,
  m,
  b,
  l,
  w,
  r,
  s = c("LC", "NT", "VU", "EN", "CR", "NE", "DD")
)
}
\arguments{
\item{t}{Time series. An integer vector of time steps, e.g. years.}

\item{j}{Number of households in the landscape. An integer vector of length 1 or \code{length(t)}.}

\item{v}{Number of human-victim incidents. An integer vector of \code{length(t)}.}

\item{m}{Number of human mortalities. An integer vector of \code{length(t)}.}

\item{b}{An integer vector of \code{length(t)}.}

\item{l}{Economic loss incidents. A list of \code{length(t)}, where each element \code{i} of that list is an integer vector of the value lost in a given incident in time \code{i}.}

\item{w}{An integer vector of 1 or \code{length(t)}.}

\item{r}{Number of wildlife-victim incidents. An integer vector of \code{length(t)}.}

\item{s}{IUCN Red List classification of the species, sub-species, or population of interest. Character vector of length 1, corresponding to one of \code{"LC"}, \code{"NT"}, \code{"VU"}, \code{"EN"}, \code{"CR"}, \code{"NE"}, or \code{"DD"}.}
}
\value{
A \code{hwcidata} object ready for analysis.
}
\description{
Data object containing data to calculate simplified HWC Index
}
\examples{

#Timeline
nyears <- 100
time_series <- 1:nyears
#landscape population
k <- 500
# average household size
ll <- 5
# number of households in landscape
j <- k/ll
#Human-victim incidents
v_mu <- 20
v_rising <- rnbinom(n = nyears, mu = v_mu + 0.2*time_series, size = v_mu/2)
#Human mortalities
m_mu <- 5
m_rising <- rnbinom(n = nyears, mu = m_mu + 0.05*time_series, size = m_mu/2) \%>\%
  if_else(v_rising < ., v_rising, .)
#Economic incidents
b_none <- rep(0, times = nyears)
#Value of economic incidents
l_none <- vector("list", length = nyears)
# Household wealth (mean)
w_bar <- 1000
#Wildlife-victim incidents
r_mu <- 5
r_rising <- rnbinom(n = nyears, mu = r_mu + 0.05*time_series, size = r_mu/2) \%>\%
  if_else(v_rising < ., v_rising, .)
# Put together into data object
sp1_data <- hwci_data(
  t = time_series,
  j = j,
  v = v_rising,
  m = m_rising,
  b = b_none,
  l = l_none,
  w = w_bar,
  r = r_rising,
  s = "CR"
)
sp1_data
}
\author{
Gerard Ryan
}
