% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwci_lm.R
\name{hwci_lm}
\alias{hwci_lm}
\title{HWC Landscape Mean Index}
\usage{
hwci_lm(...)
}
\arguments{
\item{...}{One or more \code{hwciresults} objects.}
}
\value{
A \code{hwci_lm_result} object containing landscape mean
}
\description{
Calculate Landscape Mean HWC Index
}
\examples{
allspp_lm <- hwci_lm(sp1_indices, sp2_indices, sp3_indices)
allspp_lm
}
\author{
Gerard Ryan
}
