% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hwci.R
\name{plot_hwci}
\alias{plot_hwci}
\title{HWCI plots}
\usage{
plot_hwci(data, type = c("all", "index", "sub-index", "indicator"))
}
\arguments{
\item{data}{A results object of class \code{hwciresult} or \code{hwci_lm}.}

\item{type}{Type of plot to be produced: \code{"index"}, \code{"sub-index"}, or \verb{"indicator} produce plots for single level of index suite, while default \code{"all"} produces composite of index, sub-index, and indicators. This parameter is unused where \code{"hwci_lm" \%in\% class(data)}, and a single index plot is produced in that instance.}
}
\value{
A \code{ggplot2}/\code{patchwork}-style plot.
}
\description{
Automatically generated plot types for \code{hwcresult} and \code{hwci_lm} objects.
}
\examples{
# Plot index suite
plot_hwci(sp1_indices)
# Plot only sub-indices
plot_hwci(sp2_indices, type = "sub-index")

}
