% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_hwci_data.R
\name{combine_hwci_data}
\alias{combine_hwci_data}
\title{Combine HWCI data}
\usage{
combine_hwci_data(...)
}
\arguments{
\item{...}{One or more \code{hwcidata} objects.}
}
\value{
A \code{hwcidata} object.
}
\description{
Combines several \code{hwcidata} objects into a single one (to calculate a landscape mean)
}
\author{
Gerard Ryan
}
