% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwci_data.R
\name{hwci_data}
\alias{hwci_data}
\title{HWC Index data}
\usage{
hwci_data(
  t,
  j,
  v,
  m,
  b,
  l,
  w,
  r,
  s = c("LC", "NT", "VU", "EN", "CR", "NE", "DD")
)
}
\arguments{
\item{t}{Time series. An integer vector of time steps, e.g. years.}

\item{j}{Number of households in the landscape. An integer vector of length 1 or \code{length(t)}.}

\item{v}{Number of human-victim incidents. An integer vector of \code{length(t)}.}

\item{m}{Number of human mortalities. An integer vector of \code{length(t)}.}

\item{b}{An integer vector of \code{length(t)}.}

\item{l}{Economic loss incidents. A list of \code{length(t)}, where each element \code{i} of that list is an integer vector of the value lost in a given incident in time \code{i}.}

\item{w}{An integer vector of 1 or \code{length(t)}.}

\item{r}{Number of wildlife-victim incidents. An integer vector of \code{length(t)}.}

\item{s}{IUCN Red List classification of the species, sub-species, or population of interest. Character vector of length 1, corresponding to one of \code{"LC"}, \code{"NT"}, \code{"VU"}, \code{"EN"}, \code{"CR"}, \code{"NE"}, or \code{"DD"}.}
}
\value{
A \code{hwcidata} object ready for analysis.
}
\description{
Data object containing data to calculate simplified HWC Index
}
\author{
Gerard Ryan
}
